Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main (){
-    double a,s,d;
-    scanf("%d%d%d",&a,&s,&d );
-    double n= (a+s+d)/3;
-    printf("Average: %.2d",n);
+    float a,s,d;
+    scanf("%f%f%f",&a,&s,&d );
+    float n= (a+s+d)/3;
+    printf("Average: %.2f",n);
     return 0;
 }
